+++ how-to/workspace-platform-starter/client/src/modules/integrations/workspaces/integration.ts
@@ -8,16 +8,26 @@ import type {
 	HomeSearchResult
 } from "@openfin/workspace";
 import type { Workspace, WorkspacePlatformModule } from "@openfin/workspace-platform";
+import {
+	FAVORITE_TYPE_NAME_WORKSPACE,
+	type FavoriteClient,
+	type FavoriteEntry,
+	type FavoriteInfo,
+	type FavoriteTypeNames
+} from "workspace-platform-starter/shapes/favorite-shapes";
 import type {
 	IntegrationHelpers,
 	IntegrationModule,
 	IntegrationModuleDefinition
 } from "workspace-platform-starter/shapes/integrations-shapes";
-import type { WorkspaceChangedLifecyclePayload } from "workspace-platform-starter/shapes/lifecycle-shapes";
+import type {
+	FavoriteChangedLifecyclePayload,
+	WorkspaceChangedLifecyclePayload
+} from "workspace-platform-starter/shapes/lifecycle-shapes";
 import type { Logger, LoggerCreator } from "workspace-platform-starter/shapes/logger-shapes";
 import type { ModuleDefinition } from "workspace-platform-starter/shapes/module-shapes";
 import type { ColorSchemeMode } from "workspace-platform-starter/shapes/theme-shapes";
-import { isEmpty, randomUUID } from "workspace-platform-starter/utils";
+import { isEmpty, isStringValue, randomUUID } from "workspace-platform-starter/utils";
 import type { WorkspacesSettings } from "./shapes";
 
 /**
@@ -104,6 +114,16 @@ export class WorkspacesProvider implements IntegrationModule<WorkspacesSettings>
 	private _lastResults?: HomeSearchResult[];
 
 	/**
+	 * Subscription id for theme-changed lifecycle event.
+	 */
+	private _themeChangedSubscriptionId: string | undefined;
+
+	/**
+	 * Subscription id for favorite-changed lifecycle event.
+	 */
+	private _favChangedSubscriptionId: string | undefined;
+
+	/**
 	 * Initialize the module.
 	 * @param definition The definition of the module from configuration include custom options.
 	 * @param loggerCreator For logging entries.
@@ -137,19 +157,72 @@ export class WorkspacesProvider implements IntegrationModule<WorkspacesSettings>
 							lastResult.title = payload.workspace.title;
 							lastResult.data.workspaceTitle = payload.workspace.title;
 							(lastResult.templateContent as CustomTemplate).data.title = payload.workspace.title;
+
 							this.resultAddUpdate([lastResult]);
+
+							const { favoriteClient } = await this.getFavInfo(FAVORITE_TYPE_NAME_WORKSPACE);
+							if (favoriteClient?.setSavedFavorite) {
+								const saved = await favoriteClient.getSavedFavorites(FAVORITE_TYPE_NAME_WORKSPACE);
+								const favorite = await saved?.find((f) => f.typeId === payload.id);
+								if (favorite) {
+									favorite.label = payload.workspace.title;
+									await favoriteClient.setSavedFavorite(favorite);
+								}
+							}
 						}
 					} else if (payload?.action === "delete") {
 						this.resultRemove(payload.id);
+
+						const { favoriteClient } = await this.getFavInfo(FAVORITE_TYPE_NAME_WORKSPACE);
+						if (favoriteClient?.deleteSavedFavorite) {
+							const saved = await favoriteClient.getSavedFavorites(FAVORITE_TYPE_NAME_WORKSPACE);
+							const favorite = await saved?.find((f) => f.typeId === payload.id);
+							if (favorite) {
+								await favoriteClient.deleteSavedFavorite(favorite.id);
+							}
+						}
 					}
 				}
 			);
-			this._integrationHelpers.subscribeLifecycleEvent("theme-changed", async () => {
-				if (this._integrationHelpers?.getPlatform) {
-					const platform: WorkspacePlatformModule = this._integrationHelpers.getPlatform();
-					await this.rebuildResults(platform);
+			this._themeChangedSubscriptionId = this._integrationHelpers.subscribeLifecycleEvent(
+				"theme-changed",
+				async () => {
+					if (this._integrationHelpers?.getPlatform) {
+						const platform: WorkspacePlatformModule = this._integrationHelpers.getPlatform();
+						await this.rebuildResults(platform);
+					}
 				}
-			});
+			);
+			this._favChangedSubscriptionId =
+				this._integrationHelpers.subscribeLifecycleEvent<FavoriteChangedLifecyclePayload>(
+					"favorite-changed",
+					async (_: unknown, payload?: FavoriteChangedLifecyclePayload) => {
+						if (!isEmpty(payload)) {
+							await this.updateAppFavoriteButtons(payload);
+						}
+					}
+				);
+		}
+	}
+
+	/**
+	 * Close down any resources being used by the module.
+	 * @returns Nothing.
+	 */
+	public async closedown(): Promise<void> {
+		if (this._integrationHelpers?.unsubscribeLifecycleEvent) {
+			if (isStringValue(this._themeChangedSubscriptionId)) {
+				this._integrationHelpers.unsubscribeLifecycleEvent(this._themeChangedSubscriptionId, "theme-changed");
+				this._themeChangedSubscriptionId = undefined;
+			}
+
+			if (isStringValue(this._favChangedSubscriptionId)) {
+				this._integrationHelpers.unsubscribeLifecycleEvent(
+					this._favChangedSubscriptionId,
+					"favorite-changed"
+				);
+				this._favChangedSubscriptionId = undefined;
+			}
 		}
 	}
 
@@ -159,6 +232,7 @@ export class WorkspacesProvider implements IntegrationModule<WorkspacesSettings>
 	 */
 	public async getHelpSearchEntries(): Promise<HomeSearchResult[]> {
 		if (this._integrationHelpers && this._settings) {
+			const iconFolder = await this._integrationHelpers.getCurrentIconFolder();
 			const colorScheme = await this._integrationHelpers.getCurrentColorSchemeMode();
 
 			return [
@@ -167,7 +241,7 @@ export class WorkspacesProvider implements IntegrationModule<WorkspacesSettings>
 					score: this._definition?.baseScore ?? WorkspacesProvider._DEFAULT_BASE_SCORE,
 					title: "Workspaces",
 					label: "Help",
-					icon: this._settings.images.workspace.replace("{scheme}", colorScheme as string),
+					icon: this.themeUrl(this._settings.images.workspace, iconFolder, colorScheme),
 					actions: [],
 					data: {
 						providerId: this._definition?.id
@@ -207,10 +281,12 @@ export class WorkspacesProvider implements IntegrationModule<WorkspacesSettings>
 	): Promise<HomeSearchResponse> {
 		if (this._integrationHelpers?.getPlatform && this._settings) {
 			const platform: WorkspacePlatformModule = this._integrationHelpers.getPlatform();
-			const workspaces: Workspace[] = await platform.Storage.getWorkspaces();
+			const iconFolder = await this._integrationHelpers.getCurrentIconFolder();
 			const colorScheme = await this._integrationHelpers.getCurrentColorSchemeMode();
-
 			const queryLower = query.toLowerCase();
+
+			let workspaces: Workspace[] = await platform.Storage.getWorkspaces();
+			let matchQuery = queryLower;
 
 			this._lastResponse = lastResponse;
 			this._lastQuery = queryLower;
@@ -227,7 +303,7 @@ export class WorkspacesProvider implements IntegrationModule<WorkspacesSettings>
 								key: WorkspacesProvider._ACTION_EXISTS_WORKSPACE,
 								score: this._definition?.baseScore ?? WorkspacesProvider._DEFAULT_BASE_SCORE,
 								title: `Workspace ${foundMatch.title} already exists.`,
-								icon: this._settings.images.workspace.replace("{scheme}", colorScheme as string),
+								icon: this.themeUrl(this._settings.images.workspace, iconFolder, colorScheme),
 								actions: [],
 								data: {
 									providerId: this._definition?.id,
@@ -246,7 +322,7 @@ export class WorkspacesProvider implements IntegrationModule<WorkspacesSettings>
 							key: WorkspacesProvider._ACTION_SAVE_WORKSPACE,
 							score: this._definition?.baseScore ?? WorkspacesProvider._DEFAULT_BASE_SCORE,
 							title: `Save Current Workspace as ${title}`,
-							icon: this._settings.images.workspace.replace("{scheme}", colorScheme as string),
+							icon: this.themeUrl(this._settings.images.workspace, iconFolder, colorScheme),
 							label: "Suggestion",
 							actions: [{ name: "Save Workspace", hotkey: "Enter" }],
 							data: {
@@ -262,11 +338,26 @@ export class WorkspacesProvider implements IntegrationModule<WorkspacesSettings>
 				};
 			}
 
+			const { favoriteClient, favoriteInfo } = await this.getFavInfo(FAVORITE_TYPE_NAME_WORKSPACE);
+
+			if (
+				favoriteInfo?.isEnabled &&
+				isStringValue(favoriteInfo?.command) &&
+				queryLower === favoriteInfo.command &&
+				favoriteClient
+			) {
+				const favoriteApps = await favoriteClient.getSavedFavorites(FAVORITE_TYPE_NAME_WORKSPACE);
+				const favIds = favoriteApps?.map((f) => f.typeId) ?? [];
+				workspaces = workspaces.filter((a) => favIds.includes(a.workspaceId));
+				matchQuery = "";
+			}
+
 			const workspaceResults: HomeSearchResult[] = await this.buildResults(
 				platform,
 				workspaces,
-				queryLower,
+				matchQuery,
 				options.queryMinLength,
+				iconFolder,
 				colorScheme
 			);
 
@@ -293,72 +384,98 @@ export class WorkspacesProvider implements IntegrationModule<WorkspacesSettings>
 		lastResponse: HomeSearchListenerResponse
 	): Promise<boolean> {
 		let handled = false;
-		if (this._integrationHelpers?.getPlatform && result.action.trigger === "user-action") {
-			const data: {
-				workspaceId?: string;
-				workspaceTitle?: string;
-			} = result.data;
-
-			if (data?.workspaceId) {
-				handled = true;
-
-				if (result.key === WorkspacesProvider._ACTION_SAVE_WORKSPACE) {
-					// Remove the save workspace entry
-					this.resultRemove(result.key);
-
-					const platform: WorkspacePlatformModule = this._integrationHelpers.getPlatform();
-					const snapshot = await platform.getSnapshot();
-					const currentWorkspace = await platform.getCurrentWorkspace();
-					const currentMetaData = currentWorkspace?.metadata;
-
-					const workspace = {
-						workspaceId: data.workspaceId,
-						title: data.workspaceTitle ?? "",
-						metadata: currentMetaData,
-						snapshot
-					};
-
-					await platform.Storage.saveWorkspace(workspace);
-
-					let shareEnabled: boolean = false;
-					if (this._integrationHelpers.condition) {
-						shareEnabled = await this._integrationHelpers.condition("sharing");
+		if (result.action.trigger === "user-action") {
+			if (result.action.name.endsWith("favorite") && result.data?.workspaceId) {
+				const { favoriteClient } = await this.getFavInfo(FAVORITE_TYPE_NAME_WORKSPACE);
+				if (favoriteClient) {
+					if (result.action.name.startsWith("un")) {
+						if (!isEmpty(result.data?.favoriteId) && favoriteClient.deleteSavedFavorite) {
+							await favoriteClient.deleteSavedFavorite(result.data.favoriteId);
+						}
+					} else if (favoriteClient.setSavedFavorite) {
+						await favoriteClient.setSavedFavorite({
+							id: randomUUID(),
+							type: FAVORITE_TYPE_NAME_WORKSPACE,
+							typeId: result.key,
+							label: result.title,
+							icon: this._settings?.images.workspace
+						});
 					}
-					const colorScheme = await this._integrationHelpers.getCurrentColorSchemeMode();
 
-					const savedWorkspace = await this.getWorkspaceTemplate(
-						workspace.workspaceId,
-						workspace.title,
-						shareEnabled,
-						true,
-						colorScheme
-					);
+					handled = true;
+				}
+			} else if (this._integrationHelpers?.getPlatform) {
+				const data: {
+					workspaceId?: string;
+					workspaceTitle?: string;
+				} = result.data;
 
-					// And add the new one
-					this.resultAddUpdate([savedWorkspace]);
-				} else if (result.key === WorkspacesProvider._ACTION_EXISTS_WORKSPACE) {
-					// Do nothing, the user must update the query to give it a different
-					// name which will automatically refresh the results
-				} else if (result.action.name === WorkspacesProvider._ACTION_OPEN_WORKSPACE) {
-					const platform: WorkspacePlatformModule = this._integrationHelpers.getPlatform();
-					const workspace = await platform.Storage.getWorkspace(data.workspaceId);
-					await platform.applyWorkspace(workspace);
-					// We rebuild the results here as we will now have a new current workspace
-					// and we need to change the existing one back to a standard template
-					await this.rebuildResults(platform);
-				} else if (result.action.name === WorkspacesProvider._ACTION_DELETE_WORKSPACE) {
-					const platform = this._integrationHelpers.getPlatform();
-					await platform.Storage.deleteWorkspace(data.workspaceId);
-					// Deleting the working will eventually trigger the "delete" lifecycle
-					// event which will remove it from the result list
-				} else if (
-					result.action.name === WorkspacesProvider._ACTION_SHARE_WORKSPACE &&
-					this._integrationHelpers.share
-				) {
-					await this._integrationHelpers.share({ type: "workspace", workspaceId: data.workspaceId });
-				} else {
-					handled = false;
-					this._logger?.warn(`Unrecognized action for workspace selection: ${data.workspaceId}`);
+				if (data?.workspaceId) {
+					handled = true;
+
+					if (result.key === WorkspacesProvider._ACTION_SAVE_WORKSPACE) {
+						// Remove the save workspace entry
+						this.resultRemove(result.key);
+
+						const platform: WorkspacePlatformModule = this._integrationHelpers.getPlatform();
+						const snapshot = await platform.getSnapshot();
+						const currentWorkspace = await platform.getCurrentWorkspace();
+						const currentMetaData = currentWorkspace?.metadata;
+
+						const workspace = {
+							workspaceId: data.workspaceId,
+							title: data.workspaceTitle ?? "",
+							metadata: currentMetaData,
+							snapshot
+						};
+
+						await platform.Storage.saveWorkspace(workspace);
+
+						let shareEnabled: boolean = false;
+						if (this._integrationHelpers.condition) {
+							shareEnabled = await this._integrationHelpers.condition("sharing");
+						}
+						const iconFolder = await this._integrationHelpers.getCurrentIconFolder();
+						const colorScheme = await this._integrationHelpers.getCurrentColorSchemeMode();
+
+						const { favoriteInfo } = await this.getFavInfo(FAVORITE_TYPE_NAME_WORKSPACE);
+
+						const savedWorkspace = await this.getWorkspaceTemplate(
+							workspace.workspaceId,
+							workspace.title,
+							shareEnabled,
+							true,
+							iconFolder,
+							colorScheme,
+							favoriteInfo
+						);
+
+						// And add the new one
+						this.resultAddUpdate([savedWorkspace]);
+					} else if (result.key === WorkspacesProvider._ACTION_EXISTS_WORKSPACE) {
+						// Do nothing, the user must update the query to give it a different
+						// name which will automatically refresh the results
+					} else if (result.action.name === WorkspacesProvider._ACTION_OPEN_WORKSPACE) {
+						const platform: WorkspacePlatformModule = this._integrationHelpers.getPlatform();
+						const workspace = await platform.Storage.getWorkspace(data.workspaceId);
+						await platform.applyWorkspace(workspace);
+						// We rebuild the results here as we will now have a new current workspace
+						// and we need to change the existing one back to a standard template
+						await this.rebuildResults(platform);
+					} else if (result.action.name === WorkspacesProvider._ACTION_DELETE_WORKSPACE) {
+						const platform = this._integrationHelpers.getPlatform();
+						await platform.Storage.deleteWorkspace(data.workspaceId);
+						// Deleting the working will eventually trigger the "delete" lifecycle
+						// event which will remove it from the result list
+					} else if (
+						result.action.name === WorkspacesProvider._ACTION_SHARE_WORKSPACE &&
+						this._integrationHelpers.share
+					) {
+						await this._integrationHelpers.share({ type: "workspace", workspaceId: data.workspaceId });
+					} else {
+						handled = false;
+						this._logger?.warn(`Unrecognized action for workspace selection: ${data.workspaceId}`);
+					}
 				}
 			}
 		}
@@ -372,7 +489,10 @@ export class WorkspacesProvider implements IntegrationModule<WorkspacesSettings>
 	 * @param title The title of the workspace.
 	 * @param shareEnabled Is sharing enabled.
 	 * @param isCurrent Is this the current workspace.
+	 * @param iconFolder The folder for storing icons.
 	 * @param colorScheme The current color scheme.
+	 * @param favInfo The favorites info if it is enabled.
+	 * @param favoriteId The id of the favorite.
 	 * @returns The home result.
 	 */
 	private async getWorkspaceTemplate(
@@ -380,7 +500,10 @@ export class WorkspacesProvider implements IntegrationModule<WorkspacesSettings>
 		title: string,
 		shareEnabled: boolean,
 		isCurrent: boolean,
-		colorScheme: ColorSchemeMode
+		iconFolder: string,
+		colorScheme: ColorSchemeMode,
+		favInfo: FavoriteInfo | undefined,
+		favoriteId?: string
 	): Promise<HomeSearchResult> {
 		if (this._integrationHelpers && this._settings) {
 			const actions = [
@@ -423,13 +546,30 @@ export class WorkspacesProvider implements IntegrationModule<WorkspacesSettings>
 				});
 			}
 
-			const icon = this._settings.images.workspace.replace("{scheme}", colorScheme as string);
+			const icon = this.themeUrl(this._settings.images.workspace, iconFolder, colorScheme);
+
+			const headerButtons: { icon: string; action: string }[] = [];
+
+			if (favInfo?.favoriteIcon && favInfo.unfavoriteIcon) {
+				const favoriteIcon = this.themeUrl(
+					!isEmpty(favoriteId) ? favInfo.favoriteIcon : favInfo.unfavoriteIcon,
+					iconFolder,
+					colorScheme
+				);
+				if (favoriteIcon) {
+					headerButtons.push({
+						icon: favoriteIcon,
+						action: !isEmpty(favoriteId) ? "unfavorite" : "favorite"
+					});
+				}
+			}
 
 			const layoutData = await this._integrationHelpers.templateHelpers.createLayout(
 				title,
 				icon,
 				[await this._integrationHelpers.templateHelpers.createText("instructions")],
-				actionButtons
+				actionButtons,
+				headerButtons
 			);
 
 			return {
@@ -443,7 +583,8 @@ export class WorkspacesProvider implements IntegrationModule<WorkspacesSettings>
 					providerId: this._definition?.id,
 					workspaceTitle: title,
 					workspaceId: id,
-					tags: ["workspace"]
+					tags: ["workspace"],
+					favoriteId
 				},
 				template: "Custom" as CLITemplate.Custom,
 				templateContent: {
@@ -477,7 +618,8 @@ export class WorkspacesProvider implements IntegrationModule<WorkspacesSettings>
 	 * @param platform The workspace platform.
 	 */
 	private async rebuildResults(platform: WorkspacePlatformModule): Promise<void> {
-		if (this._integrationHelpers && this._lastQuery && this._lastQueryMinLength) {
+		if (this._integrationHelpers && !isEmpty(this._lastQuery) && !isEmpty(this._lastQueryMinLength)) {
+			const iconFolder = await this._integrationHelpers.getCurrentIconFolder();
 			const colorScheme = await this._integrationHelpers.getCurrentColorSchemeMode();
 
 			const workspaces: Workspace[] = await platform.Storage.getWorkspaces();
@@ -486,6 +628,7 @@ export class WorkspacesProvider implements IntegrationModule<WorkspacesSettings>
 				workspaces,
 				this._lastQuery,
 				this._lastQueryMinLength,
+				iconFolder,
 				colorScheme
 			);
 			this.resultAddUpdate(results);
@@ -498,6 +641,7 @@ export class WorkspacesProvider implements IntegrationModule<WorkspacesSettings>
 	 * @param workspaces The list of workspaces to build the results for.
 	 * @param query The query.
 	 * @param queryMinLength The min query length.
+	 * @param iconFolder The folder for storing icons.
 	 * @param colorScheme The color scheme.
 	 * @returns The list of home search results.
 	 */
@@ -506,6 +650,7 @@ export class WorkspacesProvider implements IntegrationModule<WorkspacesSettings>
 		workspaces: Workspace[],
 		query: string,
 		queryMinLength: number,
+		iconFolder: string,
 		colorScheme: ColorSchemeMode
 	): Promise<HomeSearchResult[]> {
 		let results: HomeSearchResult[] = [];
@@ -518,21 +663,33 @@ export class WorkspacesProvider implements IntegrationModule<WorkspacesSettings>
 				shareEnabled = await this._integrationHelpers.condition("sharing");
 			}
 
+			const { favoriteClient, favoriteInfo } = await this.getFavInfo(FAVORITE_TYPE_NAME_WORKSPACE);
+			let savedFavorites: FavoriteEntry[] | undefined;
+
+			if (favoriteClient) {
+				savedFavorites = await favoriteClient.getSavedFavorites(FAVORITE_TYPE_NAME_WORKSPACE);
+			}
+
 			const wksProm = workspaces
 				.filter(
-					(pg) =>
-						query.length === 0 || (query.length >= queryMinLength && pg.title.toLowerCase().includes(query))
+					(w) =>
+						query.length === 0 || (query.length >= queryMinLength && w.title.toLowerCase().includes(query))
 				)
 				.sort((a, b) => a.title.localeCompare(b.title))
-				.map(async (ws: Workspace) =>
-					this.getWorkspaceTemplate(
+				.map(async (ws: Workspace) => {
+					const favoriteId = savedFavorites?.find((f) => f.typeId === ws.workspaceId)?.id;
+
+					return this.getWorkspaceTemplate(
 						ws.workspaceId,
 						ws.title,
 						shareEnabled,
 						currentWorkspaceId === ws.workspaceId,
-						colorScheme
-					)
-				);
+						iconFolder,
+						colorScheme,
+						favoriteInfo,
+						favoriteId
+					);
+				});
 
 			results = await Promise.all(wksProm);
 		}
@@ -573,5 +730,105 @@ export class WorkspacesProvider implements IntegrationModule<WorkspacesSettings>
 		if (this._lastResponse) {
 			this._lastResponse.revoke(id);
 		}
+	}
+
+	/**
+	 * Update the app buttons if the favorites have changed.
+	 * @param payload The payload of the favorite change.
+	 */
+	private async updateAppFavoriteButtons(payload: FavoriteChangedLifecyclePayload): Promise<void> {
+		const favorite: FavoriteEntry = payload.favorite;
+
+		if (
+			!isEmpty(this._lastResponse) &&
+			this._integrationHelpers?.getPlatform &&
+			(payload.action === "set" || payload.action === "delete") &&
+			!isEmpty(favorite) &&
+			favorite.type === FAVORITE_TYPE_NAME_WORKSPACE &&
+			this._lastResults
+		) {
+			const { favoriteInfo } = await this.getFavInfo(FAVORITE_TYPE_NAME_WORKSPACE);
+			const iconFolder = await this._integrationHelpers.getCurrentIconFolder();
+			const colorScheme = await this._integrationHelpers.getCurrentColorSchemeMode();
+
+			if (this._lastQuery === favoriteInfo?.command && payload.action === "delete") {
+				this._lastResponse.revoke(favorite.typeId);
+			} else if (this._lastResults) {
+				const lastWorkspace = this._lastResults.find((ws) => ws.key === favorite.typeId);
+
+				if (!isEmpty(lastWorkspace)) {
+					let shareEnabled: boolean = false;
+					if (this._integrationHelpers.condition) {
+						shareEnabled = await this._integrationHelpers.condition("sharing");
+					}
+
+					const platform = this._integrationHelpers.getPlatform();
+					const currentWorkspace = await platform.getCurrentWorkspace();
+					const currentWorkspaceId = currentWorkspace?.workspaceId;
+
+					const rebuilt = await this.getWorkspaceTemplate(
+						lastWorkspace.key,
+						lastWorkspace.title,
+						shareEnabled,
+						currentWorkspaceId === lastWorkspace.key,
+						iconFolder,
+						colorScheme,
+						favoriteInfo,
+						payload.action === "set" ? favorite.id : undefined
+					);
+
+					this._lastResponse.respond([rebuilt]);
+				}
+			}
+		}
+	}
+
+	/**
+	 * Get the favorite info and client if they are enabled.
+	 * @param favoriteTypeNames The type of client to get.
+	 * @returns The favorite info and client.
+	 */
+	private async getFavInfo(
+		favoriteTypeNames: FavoriteTypeNames
+	): Promise<{ favoriteClient: FavoriteClient | undefined; favoriteInfo: FavoriteInfo | undefined }> {
+		let favoriteInfo: FavoriteInfo | undefined;
+		let favoriteClient: FavoriteClient | undefined;
+
+		if ((this._definition?.data?.favoritesEnabled ?? true) && this._integrationHelpers?.getFavoriteClient) {
+			favoriteClient = await this._integrationHelpers.getFavoriteClient();
+			if (favoriteClient) {
+				favoriteInfo = favoriteClient.getInfo();
+				if (favoriteInfo.isEnabled) {
+					const isSupported =
+						isEmpty(favoriteInfo.enabledTypes) || favoriteInfo.enabledTypes.includes(favoriteTypeNames);
+					if (!isSupported) {
+						favoriteInfo = undefined;
+						favoriteClient = undefined;
+					}
+				}
+			}
+		}
+
+		return {
+			favoriteClient,
+			favoriteInfo
+		};
+	}
+
+	/**
+	 * Apply theming to an icon url.
+	 * @param url The url to theme.
+	 * @param iconFolder The icon folder.
+	 * @param colorSchemeMode The color scheme.
+	 * @returns The themed url.
+	 */
+	private themeUrl(
+		url: string | undefined,
+		iconFolder: string,
+		colorSchemeMode: ColorSchemeMode
+	): string | undefined {
+		return url
+			? url.replace(/{theme}/g, iconFolder).replace(/{scheme}/g, colorSchemeMode as string)
+			: undefined;
 	}
 }